Rename various directory to tower - Since those cases only test the tower resolver
