+++ tests/resolvers/__CASES__/tower/case01.ts
