+++ tests/resolvers/__CASES__/tower/case02.ts
