+++ tests/resolvers/__CASES__/tower/case03.ts
