+++ tests/resolvers/__CASES__/tower/case04.ts
