+++ tests/resolvers/__CASES__/tower/case05.ts
