+++ tests/resolvers/__CASES__/tower/index.ts
